inversecooking/requirements.txt
numpy==1.15.0
scipy
matplotlib
torch==0.4.1.post2
torchvision==0.2.1
nltk
Pillow
tqdm
lmdb
tensorflow
tensorboardX
flask